Updated spCoreChain.h and spCore.h Modified train and test programs for Haar and HLBP

+++ examples/boosting/trainHLBPStump.cc
@@ -62,7 +62,7 @@ int main(int argc, char* argv[])
 	TensorList *tensorList_n= manage(new TensorList()); //for validation
 
 
-	if (tensorList_p->process(p_tensor_files,target1,Tensor::Double)==false)
+	if (tensorList_p->process(p_tensor_files,target1,Tensor::Int)==false)
 	{
 		print("Error in reading +ve training patterns - Tensor list\n");
 		return 1;
@@ -70,7 +70,7 @@ int main(int argc, char* argv[])
 	pDataSet = tensorList_p->getOutput();
 
 
-	if (tensorList_n->process(n_tensor_files,target0,Tensor::Double)==false)
+	if (tensorList_n->process(n_tensor_files,target0,Tensor::Int)==false)
 	{
 		print("Error in reading -ve patterns - Tensor list\n");
 		return 1;
@@ -84,7 +84,7 @@ int main(int argc, char* argv[])
 	height = st->size(0);
 	n_examples = pDataSet->getNoExamples() + nDataSet->getNoExamples();
 	int pexamples = pDataSet->getNoExamples();
-	MemoryDataSet *m_dataset = manage(new MemoryDataSet(n_examples, Tensor::Double, true, Tensor::Short));
+	MemoryDataSet *m_dataset = manage(new MemoryDataSet(n_examples, Tensor::Int, true, Tensor::Short));
 	Tensor *example;
 
 	for (int i=0;i<pDataSet->getNoExamples();i++)
